rename the sqldb_client to nosqldb_client
