+++ app/clients/nosqldb_client.py
